+++ src/server/game/Entities/Unit/Unit.h
@@ -2030,11 +2030,11 @@ public:
     int32 GetCreatePowers(Powers power) const;
     float GetPosStat(Stats stat) const
     {
-        return GetFloatValue((uint16)UNIT_FIELD_STAT_POS_BUFF + (uint16)stat);
+        return GetFloatValue(UNIT_FIELD_STAT_POS_BUFF + stat);
     }
     float GetNegStat(Stats stat) const
     {
-        return GetFloatValue((uint16)UNIT_FIELD_STAT_NEG_BUFF + (uint16)stat);
+        return GetFloatValue(UNIT_FIELD_STAT_NEG_BUFF + stat);
     }
     float GetCreateStat(Stats stat) const
     {
